Added checks in except branch protection and githubpages - works in app.js - will move to index tomorrow and format branch protection
