--- scanners/github-octokit-checks/src/main-language-check.js
@@ -1,7 +1,0 @@
-export async function getMainLanguage(repoDetails) {
-    if (repoDetails.language) {
-        return({"mainLanguage": repoDetails.language})
-    } else {
-        return({"mainLanguage": null})
-    }
-}
