+++ scanners/github-octokit-checks/src/octokit-check-context.js
@@ -1,0 +1,9 @@
+
+export class RepoChecker {
+    async makeOctokitRequest(processor, branchName='main'){
+        return processor.makeOctokitRequest(branchName)
+    }
+    formatResponse(processor){
+        return processor.formatResponse()
+    }
+}
